requirements.txt
checks-superstaq~=0.5.64
cirq-superstaq~=0.5.64
magic-state-cultivation @ git+https://git@github.com/Infleqtion/magic-state-cultivation.git@toberoi/add_fold_traversal
openfermion>=1.7.1
pygridsynth>=1.2.0
qiskit-superstaq~=0.5.64
tqdm~=4.67
stimcirq~=1.15
